refactor: Rename embedders/system_api.rs to embedders/linker.rs
