+++ rs/embedders/src/wasmtime_embedder/linker.rs
